requirements.txt
# Smart File Organizer 依赖包

# AI 工作流框架
langgraph>=0.0.40

# LangChain 和 OpenAI 集成
langchain-openai>=0.1.0
langchain-core>=0.2.0

# 数据验证
pydantic>=2.5.0

# 文档处理
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.21

# 配置文件解析
pyyaml>=6.0.1

# 日志工具（更优雅的日志输出）
loguru>=0.7.2

# 文件系统监听（用于实时监控文件变化）
watchdog>=4.0.0

# 环境变量管理
python-dotenv>=1.0.0

# Web UI (Step 7 Phase 2)
streamlit>=1.28.0

# OCR 和图像处理 (智能 PDF 识别增强)
rapidocr-onnxruntime>=1.3.0
pdf2image>=1.16.0
Pillow>=10.0.0

# 注意：pdf2image 需要系统安装 poppler
# Windows: 下载 poppler 并配置 PATH
# macOS: brew install poppler
# Linux: apt-get install poppler-utils

# Memory System (审批日志和偏好学习)
sqlalchemy>=2.0.0

